1. use type bias; 2. use mse: (Etot-Epre)/atoms; 3. add least_square; 4. optimize memory

+++ src/main_nep_adam/gradients.cuh
@@ -27,26 +27,26 @@ struct Gradients {
     grad_wb_sum.resize(number_of_variables_ann);
   }
 
-  GPU_Vector<float> E_c;         // partial energy / partial C
-  GPU_Vector<float> F_c_x;       // partial force / partial C
-  GPU_Vector<float> F_c_y;       // partial force / partial C
-  GPU_Vector<float> F_c_z;       // partial force / partial C
-  GPU_Vector<float> V_c_xx;      // partial virial_xx / partial C
-  GPU_Vector<float> V_c_yy;      // partial virial_yy / partial C
-  GPU_Vector<float> V_c_zz;      // partial virial_zz / partial C
-  GPU_Vector<float> V_c_xy;      // partial virial_xy / partial C
-  GPU_Vector<float> V_c_yz;      // partial virial_yz / partial C
-  GPU_Vector<float> V_c_zx;      // partial virial_zx / partial C
-  GPU_Vector<float> grad_c_sum;  // sum of energy, force, virial w.r.t. C
-  GPU_Vector<float> E_wb_grad;      // energy w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> F_wb_grad_x;      // force w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> F_wb_grad_y;      // force w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> F_wb_grad_z;      // force w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_xx;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_yy;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_zz;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_xy;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_yz;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> V_wb_grad_zx;      // virial w.r.t. w0, b0, w1, b1
-  GPU_Vector<float> grad_wb_sum;      // sum energy, force, virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> E_c;         // partial energy / partial C
+  GPU_Vector<double> F_c_x;       // partial force / partial C
+  GPU_Vector<double> F_c_y;       // partial force / partial C
+  GPU_Vector<double> F_c_z;       // partial force / partial C
+  GPU_Vector<double> V_c_xx;      // partial virial_xx / partial C
+  GPU_Vector<double> V_c_yy;      // partial virial_yy / partial C
+  GPU_Vector<double> V_c_zz;      // partial virial_zz / partial C
+  GPU_Vector<double> V_c_xy;      // partial virial_xy / partial C
+  GPU_Vector<double> V_c_yz;      // partial virial_yz / partial C
+  GPU_Vector<double> V_c_zx;      // partial virial_zx / partial C
+  GPU_Vector<double> grad_c_sum;  // sum of energy, force, virial w.r.t. C
+  GPU_Vector<double> E_wb_grad;      // energy w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> F_wb_grad_x;      // force w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> F_wb_grad_y;      // force w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> F_wb_grad_z;      // force w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_xx;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_yy;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_zz;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_xy;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_yz;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> V_wb_grad_zx;      // virial w.r.t. w0, b0, w1, b1
+  GPU_Vector<double> grad_wb_sum;      // sum energy, force, virial w.r.t. w0, b0, w1, b1
 };
